soporte/requirements.txt
Django==4.0.2
django-import-export==2.7.1
openpyxl==3.0.9
wfastcgi==3.0.0
django-mssql-backend==2.8.1
djangorestframework==3.13.1
requests==2.27.1